relationship one to many done

+++ full stack/backend/route/Note.route.js
@@ -19,9 +19,10 @@ try {
 // read
 
 NoteRouter.get("/", async(req,res)=>{
+    const { authorID } =req.body; 
     try {
-        const user = await NoteModel.find();
-        res.status(200).send(user);
+        const note = await NoteModel.find({authorID:authorID});
+        res.status(200).send(note);
     } catch (err) {
         res.status(400).send({"err":err.message});
     }
@@ -30,11 +31,17 @@ NoteRouter.get("/", async(req,res)=>{
 
 // update
 
-NoteRouter.patch("/update/:id", async(req,res)=>{
-    const {id} = req.params;
+NoteRouter.patch("/update/:Id", async(req,res)=>{
+    const {Id} = req.params;
+    const note = await NoteModel.findOne({_id:Id})
     try {
-        await NoteModel.findByIdAndUpdate({_id:id},req.body);
-        res.status(200).send({"msg":"Note Updated Successfully"});
+        if(note.authorID==req.body.authorID){
+            await NoteModel.findByIdAndUpdate({_id:Id},req.body);
+            res.status(200).send({"msg":"Note Updated Successfully"});
+        }else{
+           res.status(200).send({"msg":"You are not authorized to update this note."})
+        }
+       
     } catch (err) {
         res.status(400).send({"err":err.message});
     }
@@ -45,9 +52,16 @@ NoteRouter.patch("/update/:id", async(req,res)=>{
 
 NoteRouter.delete("/delete/:id", async(req,res)=>{
     const {id} = req.params;
+    const note =await NoteModel.findOne({_id:id});
     try {
-        await NoteModel.findByIdAndDelete({_id:id});
-        res.status(200).send({"msg":"Note Deleted Successfully"});
+        if(req.body.authorID==note.authorID){
+            await NoteModel.findByIdAndDelete({_id:id});
+            res.status(200).send({"msg":"Note Deleted Successfully"});
+        }else{
+            res.status(200).send({"msg":"You are not authorised to delete this note."});
+
+        }
+        
     } catch (err) {
         res.status(400).send({"err":err.message});
     }
